Improve hierarchy of files
